Department
Upadate Department delete by ID and all for not throw error 500

+++ src/main/java/com/project/professor/allocation/service/DepartmentService.java
@@ -8,16 +8,20 @@ import org.springframework.stereotype.Service;
 import com.project.professor.allocation.entity.Department;
 import com.project.professor.allocation.entity.Department;
 import com.project.professor.allocation.repository.DepartmentRepository;
+import com.project.professor.allocation.repository.ProfessorRepository;
+import com.project.professor.allocation.service.exception.AllocatedExistsException;
 import com.project.professor.allocation.service.exception.EntityNotFindException;
 
 @Service
 public class DepartmentService {
 
 	private final DepartmentRepository departmentRepository;
+	private final ProfessorRepository professorRepository;
 
-	public DepartmentService(DepartmentRepository departmentRepository) {
+	public DepartmentService(DepartmentRepository departmentRepository, ProfessorRepository professorRepository) {
 		super();
 		this.departmentRepository = departmentRepository;
+		this.professorRepository = professorRepository;
 	}
 
 	public List<Department> findByNameContaining(String name) {
@@ -46,15 +50,27 @@ public class DepartmentService {
 		}
 	}
 
-	public void deleteById(Long id) throws EntityNotFindException {
+	public void deleteById(Long id) throws EntityNotFindException, AllocatedExistsException {
 		if (id != null && departmentRepository.existsById(id)) {
-			departmentRepository.deleteById(id);
+			if(professorRepository.findByDepartmentId(id) == null) {
+				departmentRepository.deleteById(id);	
+			}else {
+				throw new AllocatedExistsException("This department have professor");
+			}
 		} else {
 			throw new EntityNotFindException("Department ID doesnt exists");
 		}
 	}
 
-	public void deleteAll() {
+	public void deleteAll() throws AllocatedExistsException {
+		
+		List<Department> departments = findAll();
+		
+		for (Department department : departments) {
+			if(professorRepository.findByDepartmentId(department.getId()) != null) {
+				throw new AllocatedExistsException("Department have professor(s)");	
+			}
+		}
 		departmentRepository.deleteAll();
 	}
